Reorganize smt-related features
