+++ src/CMakeLists.txt
@@ -13,6 +13,7 @@ add_subdirectory(ila-mngr)
 add_subdirectory(mcm)
 add_subdirectory(target-json)
 add_subdirectory(target-sc)
+add_subdirectory(target-smt)
 add_subdirectory(util)
 add_subdirectory(verilog-in)
 add_subdirectory(verilog-out)
