+++ src/ila-mngr/CMakeLists.txt
@@ -12,7 +12,6 @@ target_sources(${ILANG_LIB_NAME} PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/u_abs_knob.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/u_rewrite_expr.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/u_rewrite_ila.cc
-  ${CMAKE_CURRENT_SOURCE_DIR}/u_smt_switch.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/u_unroller.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/v_eq_check.cc
   ${CMAKE_CURRENT_SOURCE_DIR}/v_eq_check_crr.cc
