+++ src/ila/validate_model.cc
@@ -3,9 +3,10 @@
 
 #include <ilang/ila/validate_model.h>
 
-#include "z3++.h"
+#include <z3++.h>
+
 #include <ilang/ila/ast_fuse.h>
-#include <ilang/ila/z3_expr_adapter.h>
+#include <ilang/target-smt/z3_expr_adapter.h>
 #include <ilang/util/log.h>
 
 namespace ilang {
