+++ test/t_smt_switch_itf.cc
@@ -4,8 +4,8 @@
 #include <smt-switch/boolector_factory.h>
 #include <smt-switch/smt.h>
 
-#include <ilang/ila-mngr/u_smt_switch.h>
 #include <ilang/ilang++.h>
+#include <ilang/target-smt/smt_switch_itf.h>
 
 #include "unit-include/util.h"
